requirements.txt
langchain>=0.0.350,<0.1.0
langchain-google-genai==0.0.7
pandas==2.1.4
sqlalchemy==2.0.25
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain-community==0.0.17
langchain-core>=0.1.16,<0.2.0
langchain-experimental==0.0.47
streamlit==1.32.0
atlassian-python-api>=3.30.0
